pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "experiments"
version = "0.0.1"
authors = [
    { name = "Richard Dong", email = "richard@tristero.xyz" },
]
description = "Personal experiments"
readme = "README.md"
requires-python = ">=3.12.6"
dependencies = [
    "google-cloud-secret-manager",
    "python-telegram-bot",
]

[project.optional-dependencies]
dev = ["build", "twine"]

[project.scripts]
main = "experiments.__main__:main"
